Use anchor tag for normal links in chat for more native experience.
Using anchor tags has the advantage that the browser displays it as a
normal link, like all links on the internet. It lets you right click
the link, copy the URL to clipboard, open in tab or window, or
incognito window, etc. It's better to use the HTML standard than try to
mimic such behavior with custom JavaScript when possible.
Note that this is only used for regular links that are posted in chat;
the internal WebIRC javascript-based links (like joining a channel,
etc.) remain as they are.
Use same class name for chat log links. Different tags can be selected
in css.